move to common_config
